backend/requirements.txt
# Hotel Voice AI Concierge - Python Dependencies
# Core requirements for the voice-enabled hotel room service system


# Web Framework & API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic
pydantic-settings==2.1.0


# Database & ORM  
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1


# Voice AI Pipeline (Pipecat)
pipecat-ai-flows
pipecat-ai[webrtc,silero,deepgram,openai,cartesia,runner,daily,google,anthropic,groq]>=0.0.77

# Speech Services
# Note: Soniox requires separate installation from their repository
deepgram-sdk==3.2.7
cartesia==1.0.5


# LLM Services
openai==1.3.8
requests==2.31.0


# Voice Transport (Daily.co)
# Note: Daily transport is included in pipecat-ai


# Audio Processing
numpy==1.26.4

# this causes some issues with mac osx 
# in that case have used "brew install portaudio" then after that this command "pip install --global-option='build_ext' --global-option="-I$(brew --prefix)/include" --global-option="-L$(brew --prefix)/lib" pyaudio"
pyaudio==0.2.11


# Environment & Configuration
python-dotenv==1.0.0
loguru==0.7.2


# Development & Utilities
aiohttp==3.10.3
asyncio-mqtt==0.16.1


# Data Processing
pandas==2.1.4
jinja2==3.1.2


# Security - FIXED VERSION
cryptography==42.0.8
passlib[bcrypt]==1.7.4


# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2


# CORS Support
python-multipart==0.0.6


# Email validation
email-validator==2.1.0


# Date/Time utilities
python-dateutil==2.8.2
